requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
httpx==0.26.0
pydantic==2.7.0
python-dotenv==1.1.0

# dev
pytest==7.4.0
pytest-asyncio==0.22.0
black==24.3.0
ruff==0.6.9
